выполнение задания 6.2
